requirements.txt
watchdog>=4.0.0
paramiko>=3.4.0
pystray>=0.19.5
Pillow>=10.2.0
